requirements.txt
cffi==1.13.2
cryptography==2.8
dj-database-url==0.5.0
Django==2.2.7
django-heroku==0.3.1
django-rest-knox==4.1.0
django-webpack-loader==0.6.0
djangorestframework==3.10.3
gunicorn==20.0.4
Pillow==6.2.1
psycopg2==2.8.4
pycparser==2.19
python-dotenv==0.10.5
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
whitenoise==5.0.1
